restructuring messaging UI components
