+++ src/components/messages/FriendRequest.vue
